Pain
Apparently this game's code is so bumfuzzled that these don't work even though I tested t hem and they worked. It would be EXTREMELY HELPFUL if someone helped me comment them out. I spent an hour looking for an alternative, but sadly it seems that ai percentage modifiers work about as well as the global economy.