Fix
I can't figure out if its the events, decision, something in the code not being read - but PA doesn't accurately implement the checks for this decision or the events. No idea what is wrong or how to stop it - though maybe this small change helps it stop happening in other places.