Operation Nehalennia 1/??????
Owing to increased performance with the modified engine, I have opted tentatively to add in new religions and cultures. This all pending assistance of course. The 1946 start date will be the only one modified, for now, though in the future perhaps the others will follow. Currently there isn't much functionality to the new stuff beyond flavor, however I hope to add in more in the future alongside integration into existing and new/future events! The next steps are to add in variants of Folk Christianity, a tall order given that such things are not really recorded during a census - though something that can perhaps add a tiny bit of flavor and varience to nations.